Created CMakeLists.txt file and fixed include paths and namespaces
