+++ Rendering/Effects/BoxEffect.cpp
@@ -18,14 +18,14 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 #include "BoxEffect.h"
 
 
-#include "../UserInterface/ViewPanel/RenderResources.h"
+#include "../Resources/RenderResources.h"
 #include <BlueFramework/Rasterizer/vertex.h>
 #include <BlueFramework/Engine/Mesh/geometryGeneration.h>
 
 #include <tuple>
 
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_BEGIN
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_BEGIN
 
 
 BoxEffect::BoxEffect(buw::IRenderSystem * renderSystem, buw::ReferenceCounted<buw::IViewport> viewport, buw::ReferenceCounted<buw::ITexture2D> depthStencilMSAA, buw::ReferenceCounted<buw::IConstantBuffer> worldBuffer, buw::ReferenceCounted<buw::IConstantBuffer> viewportBuffer)
@@ -103,4 +103,4 @@ void BoxEffect::v_render()
 	}
 }
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_END
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_END
