+++ Rendering/Effects/BoxEffect.h
@@ -27,7 +27,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 #include <map>
 #include <tuple>
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_BEGIN
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_BEGIN
 
 class BoxEffect : public buw::Effect {
 public:	
@@ -66,6 +66,6 @@ private:
 	bool bShow_ = false;
 };
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_END
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_END
 
 #endif //end define OpenInfraPlatform_UserInterface_BoxEffect_F594ABC3_4E77_4E7B_BB89_2A31CD33156A_h
