+++ Rendering/Effects/GradientClearEffect.cpp
@@ -17,7 +17,7 @@
 
 #include "GradientClearEffect.h"
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_BEGIN
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_BEGIN
 
 GradientClearEffect::GradientClearEffect(buw::IRenderSystem* renderSystem, buw::ReferenceCounted<buw::IViewport> viewport) : Effect(renderSystem), viewport_(viewport) {
 
@@ -89,4 +89,4 @@ void GradientClearEffect::v_init() {
 	}
 }
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_END
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_END
