+++ Rendering/Effects/SkyboxEffect.h
@@ -27,7 +27,7 @@
 #include <buw.Rasterizer.h>
 #include <map>
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_BEGIN
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_BEGIN
 
 class SkyboxEffect : public buw::Effect {
 public:
@@ -51,6 +51,6 @@ private:
     buw::ReferenceCounted<buw::IIndexBuffer> indexBuffer_;
 };
 
-OIP_NAMESPACE_OPENINFRAPLATFORM_UI_END
+OIP_NAMESPACE_OPENINFRAPLATFORM_RENDERING_END
 
 #endif // end define BlueFramework_UserInterface_SkyboxEffect_34f22a43_6c1f_4b9f_99f5_ca188f4c0343_h
